// frontend-legacy/src/client/types.gen.ts
export type ContentFormat = 'text' | 'html' | 'markdown' | 'json';
// ...
export type StoryNodePublic = {
    title: string;
    content?: string;
    content_format?: (ContentFormat | null);
    node_type?: (string | null);
    is_start_node?: boolean;
    is_end_node?: boolean;
    id: string;
    story_id: string;
    story_version: number;
    created_at: string;
    updated_at: string;
};
// ...
export type StoryPublic = {
    title: string;
    content_format?: (ContentFormat | null);
    description?: (string | null);
    is_published: boolean;
    id: string;
    owner_id: string;
    current_version: number;
    published_version: (number | null);
    created_at: string;
    updated_at: string;
};